[refactor] 調整 controller 與 route 檔案結構至 user 子目錄
