+++ controllers/user/userControllers.js
@@ -1,6 +1,6 @@
-const successHandler = require("../services/successHandler");
-const appError = require("../services/appError");
-const User = require("../models/user");
+const successHandler = require("../../services/successHandler");
+const appError = require("../../services/appError");
+const User = require("../../models/user");
 
 const UserControllers = {
   async getUsers(req, res, next) {
