+++ routes/user/postRoutes.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const router = express.Router();
-const PostControllers = require("../controllers/PostControllers");
-const errorAsyncHandler = require("../services/errorAsyncHandler");
+const PostControllers = require("../../controllers/user/postControllers");
+const errorAsyncHandler = require("../../services/errorAsyncHandler");
 
 router.get("/posts", errorAsyncHandler(PostControllers.getPosts));
