+++ routes/user/userRoutes.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const router = express.Router();
-const UserControllers = require("../controllers/UserControllers");
-const errorAsyncHandler = require("../services/errorAsyncHandler");
+const UserControllers = require("../../controllers/user/userControllers");
+const errorAsyncHandler = require("../../services/errorAsyncHandler");
 
 router.get("/users", errorAsyncHandler(UserControllers.getUsers));
